slidenotes-backend/convertPPTXToText/requirements.txt
requests
python-pptx
boto3
nltk
